pyproject.toml
[build-system]
requires = ["uv_build>=0.7.10,<0.8.0"]
build-backend = "uv_build"

[project]
name = "reality-defender-slack-app"
version = "0.1.0"
description = "An app that integrates Slack with the Reality Defender platform."
authors = [
    { name = "Reality Defender", email = "support@realitydefender.ai" }
]
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.12"
keywords = ["slack", "app", "deepfake", "async"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications :: Chat",
]

dependencies = [
    "aiohttp>=3.9.0",
    "requests>=2.32.4",
    "slack-bolt>=1.23.0",
    "pydantic>=2.11.4,<3.0.0",
    "python-dotenv>=1.0.0",
    "realitydefender>=0.1.2"
]

[dependency-groups]
dev = [
    "types-requests>=2.32.4",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "ruff>=0.11.3",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
]

lint = [
    "ruff>=0.11.3",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/reality-defender/slack-app"
Repository = "https://github.com/reality-defender/slack-app"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"
markers = [
    "asyncio: mark a test as an asyncio coroutine"
]
asyncio_default_fixture_loop_scope = "function"

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true
allow_redefinition = true

[[tool.mypy.overrides]]
module = ["slack_bolt.*", "aiohttp.*"]
ignore_missing_imports = true